SDI-345: ♻️  Create select hospital service
