+++ server/routes/addPatient/index.ts
@@ -5,15 +5,15 @@ import PrisonerSearchService from '../../services/prisonerSearchService'
 import PrisonerSearchRoutes from './prisonerSearch'
 import PrisonerSelectRoutes from './prisonerSelect'
 import authorisationMiddleware from '../../middleware/authorisationMiddleware'
-import MovePrisonerService from '../../services/movePrisonerService'
-import MovePrisonerRoutes from './hospitalSelect'
+import HospitalSearchService from '../../services/hospitalSearchService'
+import HospitalSelectRoutes from './hospitalSelect'
 
 export default function addPrisonerRoutes({
-  movePrisonerService,
+  hospitalSearchService,
   prisonerSearchService,
 }: {
   prisonerSearchService: PrisonerSearchService
-  movePrisonerService: MovePrisonerService
+  hospitalSearchService: HospitalSearchService
 }): Router {
   const router = express.Router({ mergeParams: true })
   router.use(authorisationMiddleware(true, ['RESTRICTED_PATIENT_MIGRATION']))
@@ -23,7 +23,7 @@ export default function addPrisonerRoutes({
 
   const prisonerSearch = new PrisonerSearchRoutes()
   const prisonerSelect = new PrisonerSelectRoutes(prisonerSearchService)
-  const hospitalSelect = new MovePrisonerRoutes(movePrisonerService, prisonerSearchService)
+  const hospitalSelect = new HospitalSelectRoutes(hospitalSearchService, prisonerSearchService)
 
   get('/search-for-prisoner', prisonerSearch.view)
   post('/search-for-prisoner', prisonerSearch.submit)
