+++ server/routes/movePrisoner/hospitalSelect.ts
@@ -1,11 +1,11 @@
 import HospitalSelectRoutes from '../selectHospital/hospitalSelect'
-import MovePrisonerService from '../../services/movePrisonerService'
+import HospitalSearchService from '../../services/hospitalSearchService'
 import PrisonerSearchService from '../../services/prisonerSearchService'
 
 export default class MovePrisonerHospitalSelectRoutes extends HospitalSelectRoutes {
-  constructor(movePrisonerService: MovePrisonerService, prisonerSearchService: PrisonerSearchService) {
+  constructor(hospitalSearchService: HospitalSearchService, prisonerSearchService: PrisonerSearchService) {
     super(
-      movePrisonerService,
+      hospitalSearchService,
       prisonerSearchService,
       '/move-to-hospital/confirm-move',
       'pages/movePrisoner/movePrisonerSelectHospital'
