+++ server/services/index.ts
@@ -6,11 +6,13 @@ import HmppsAuthClient from '../data/hmppsAuthClient'
 
 import TokenStore from '../data/tokenStore'
 import RemoveRestrictedPatientService from './removeRestrictedPatientService'
+import HospitalSearchService from './hospitalSearchService'
 
 const hmppsAuthClient = new HmppsAuthClient(new TokenStore())
 const userService = new UserService(hmppsAuthClient)
 const prisonerSearchService = new PrisonerSearchService(hmppsAuthClient)
-const movePrisonerService = new MovePrisonerService(hmppsAuthClient)
+const movePrisonerService = new MovePrisonerService()
+const hospitalSearchService = new HospitalSearchService()
 const restrictedPatientSearchService = new RestrictedPatientSearchService()
 const removeRestrictedPatientService = new RemoveRestrictedPatientService(hmppsAuthClient)
 
@@ -20,6 +22,7 @@ export const services = {
   movePrisonerService,
   restrictedPatientSearchService,
   removeRestrictedPatientService,
+  hospitalSearchService,
 }
 
 export type Services = typeof services
